+++ sdk/react/src/index.ts
@@ -11,6 +11,7 @@ export * from "./hooks/useConnection.js";
 export * from "./hooks/useMcpConnections.js";
 export * from "./hooks/useMcp.js";
 export * from "./hooks/useSkills.js";
+export * from "./hooks/useRcpSources.js";
 export * from "./hooks/useWorkflows.js";
 export * from "./hooks/useWorkflow.js";
 export * from "./hooks/useWorkflowStream.js";